+++ collect-webapp/frontend/src/datamanagement/components/recordeditor/MultipleFieldset.js
@@ -4,6 +4,7 @@ import { Button, Input } from 'reactstrap'
 import FormItems from './FormItems'
 import EntityCollectionComponent from './EntityCollectionComponent'
 import TabSetContent from './TabSetContent'
+import DeleteNodeButton from './DeleteNodeButton'
 
 export default class MultipleFieldset extends EntityCollectionComponent {
   constructor() {
@@ -74,11 +75,7 @@ export default class MultipleFieldset extends EntityCollectionComponent {
         <Button color="success" onClick={this.handleNewButtonClick}>
           New
         </Button>
-        {selectedEntity && (
-          <Button color="danger" onClick={this.handleDeleteButtonClick}>
-            <span className="fa fa-trash" />
-          </Button>
-        )}
+        {selectedEntity && <DeleteNodeButton onClick={this.handleDeleteButtonClick} />}
         {selectedEntity && (
           <>
             <FormItems itemDefs={itemDef.items} parentEntity={selectedEntity} />