+++ collect-webapp/frontend/src/model/ui/MultipleFieldDefinition.js
@@ -1,0 +1,3 @@
+import { FieldDefinition } from './FieldDefinition'
+
+export class MultipleFieldDefinition extends FieldDefinition {}