+++ CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [3.26.24] - NOT RELEASED
+### Added
 - Removed Planet API key field from CE surveys
 - Activated Planet imagery by default when creating new project
-### Added
-- Data entry new UI: support multiple entities with table layout / coordinate attributes
+- Data entry: support column groupings in table (single entities inside multiple entities with table layout)
 
 ## [3.26.23] - 2020-10-14
 ### Fixed